Added confirm dialog for delete

+++ src/components/Content/Transactions/useTransactionDetailsDialogActions.ts
@@ -10,6 +10,8 @@ import {
 	useUpdateTransactions
 } from '../../../ajaxapi/query/TransactionQueries';
 import { pipe } from 'fp-ts/es6/function';
+import { useContext } from 'react';
+import { ConfirmDialogContext } from '../../UI/ConfirmDialog/ConfirmDialogProvider';
 
 interface TransactionDetailsDialogState {
 	readonly selectedTransaction: OptionT<TransactionResponse>;
@@ -25,6 +27,7 @@ interface TransactionDetailsDialogActions {
 
 export const useTransactionDetailsDialogActions =
 	(): TransactionDetailsDialogActions => {
+		const { newConfirmDialog } = useContext(ConfirmDialogContext);
 		const [detailsDialogState, setDetailsDialogState] =
 			useImmer<TransactionDetailsDialogState>({
 				selectedTransaction: Option.none
@@ -54,9 +57,14 @@ export const useTransactionDetailsDialogActions =
 					(id) => [id]
 				)
 			);
-			deleteTransactionsMutate({
-				idsToDelete
-			});
+			newConfirmDialog(
+				'Delete Transaction',
+				'Are you sure you want to delete this transaction?',
+				() =>
+					deleteTransactionsMutate({
+						idsToDelete
+					})
+			);
 		};
 
 		return {
